refactor: split CLI command families
